requirements.txt
annotated-types==0.7.0
anyio==4.12.1
backports.asyncio.runner==1.2.0
Bottleneck==1.6.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.4
distro==1.9.0
docstring_parser==0.17.0
exceptiongroup==1.3.1
filetype==1.2.0
google-api-core==2.29.0
google-auth==2.48.0
google-cloud-aiplatform==1.136.0
google-cloud-bigquery==3.40.0
google-cloud-core==2.5.0
google-cloud-resource-manager==1.16.0
google-cloud-storage==3.9.0
google-cloud-vectorsearch==0.4.0
google-crc32c==1.8.0
google-genai==1.62.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpc-google-iam-v1==0.14.3
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
iniconfig==2.3.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.9
langchain-core==1.2.9
langchain-google-genai==4.2.0
langchain-google-vertexai==3.2.2
langchain-tests==1.1.4
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.4
langsmith==0.7.1
markdown-it-py==4.0.0
mdurl==0.1.2
numexpr==2.14.1
numpy==2.2.6
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pluggy==1.6.0
proto-plus==1.27.1
protobuf==6.33.5
py-cpuinfo==9.0.0
pyarrow==22.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==8.4.2
pytest-asyncio==1.3.0
pytest-benchmark==5.2.3
pytest-codspeed==4.3.0
pytest-recording==0.13.4
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
syrupy==4.9.1
tenacity==9.1.4
tomli==2.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.0
validators==0.35.0
vcrpy==8.1.1
websockets==15.0.1
wrapt==2.1.1
xxhash==3.6.0
zstandard==0.25.0